+++ src/Resources/config/invalidation_listener.php
@@ -4,7 +4,6 @@ namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
 return static function (ContainerConfigurator $container) {
     $services = $container->services();
-    $parameters = $container->parameters();
 
     $services->set('fos_http_cache.event_listener.invalidation', \FOS\HttpCacheBundle\EventListener\InvalidationListener::class)
         ->args([
@@ -12,6 +11,7 @@ return static function (ContainerConfigurator $container) {
             service('router'),
             service('fos_http_cache.rule_matcher.must_invalidate'),
             service('fos_http_cache.invalidation.expression_language')->ignoreOnInvalid(),
+            '%fos_http_cache.invalidation.generate_url_type%',
         ])
         ->tag('kernel.event_subscriber');
 };